Don't give the user action to re-run fetch transactions
We gave the user a closure that lets the user re-run fetch actions. As
request ID is in the closure and would be retried with same ID every
time, re-running this action would always give the user `Left mempty`:
this seems like a very useless thing to give to the user.

Instead, we ask the user for callbacks on how to manipulate responses.
The user is also able to instantly indicate whether or not they want
to handle delayed responses. I have not done it in this commit but
this now allows us to reduce amount of work the server has to do if
the user only cares about instant responses.

There's a new implementation (and type change) of `traverse` as a
result: there seemed to be no point of any atomic transactions: each
request can only have a single result so whether we were running the
whole transaction atomically or not, it didn't matter: it was just a
fairly strange way to gather results. Instead, we'll simply use
`async`'s `forConcurrently` which will gather results as they come in:
we could use a simple `traverse` of course but in presence of delayed
responses, this seems inefficient. This is up for discussion – perhaps
this library function shouldn't exist anymore at all: it is trivially
implementable by the user.