Simplify request resource handling.
Previously, we had cases handling things like "what happens if request
we're waiting on suddenly disappears" &c. This was only possible if
the user happened to get a hold of a re-runnable `fetch` action for
the same request ID. One of the previous commits made this impossible
so these cases were no longer possible, but we still had to handle
them due to how our resource tracking worked.

This meant that it was wise to rework request resource tracking such
that request we're working with is always present and nothing can
destroy it except the whole fetch action finishing.

This also allows for much more precise handling of the request
life-time in case of exceptions (like user cancellations). Previously
we were testing if request still seems alive and cancelling requests
then. This makes little sense with current approach as we now
precisely know when request resource is alive and when it's dead: it's
always alive until the user is finished.

There was also the case of reply handler which peeked into the request
map and updated shared state. It still does it, but it's done in a
much more responsive way than before: previously, if a reply to a
request came, we would block the whole request for the whole duration
of handling the request. In a situation with a lot of requests, we
would became much less responsive than we'd have to be: now the lock
is taken only for a brief while after which we manipulate the value
inside.

As internally the concurrent resource map uses MVar over TVar, the
sytem should be much more responsive even if granularity of locking
wasn't improved.