De-duplicate response type
This difference only served as a run-time assertion that we only send
delayed response after signalling it will be delayed: it was not very
useful unless we have a glaring implementation error in `respond`.
It'd be much more useful to have tests for this rather than assert
that should never fail.