Don't check for duplicate requests, non-racey thread map
The only way to send requests is from `fetch` which enforces that each
request ID is unique.

Further, there was an easy possibility of two request handlers
over-writing one another in the thread map IORef which could leave a
thread hanging around indefinitely – not good. This bug would not come
up with the way the code is right now as the requests are handled in
single-thread, sequentially. However, better safe than sorry as this
limitation may be lifted in the future: it seems quite bad for
through-put and latency to have it.